refactoring models. still have to fix env.py
