+++ backend/app/api/routes/inquiries.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, HTTPException
 
 import app.crud.inquiry as crud
 from app.api.deps import SessionDep
-from app.models import Inquiry, InquiryCreate, InquiryPublic
+from app.models.inquiry import Inquiry, InquiryCreate, InquiryPublic
 
 router = APIRouter()
