+++ backend/app/crud/user.py
@@ -3,7 +3,7 @@ from typing import Any
 from sqlmodel import Session, select
 
 from app.core.security import get_password_hash, verify_password
-from app.models import (
+from app.models.user import (
     User,
     UserCreate,
     UserUpdate,
