+++ backend/app/tests/api/routes/test_users.py
@@ -7,7 +7,7 @@ from sqlmodel import Session, select
 import app.crud.user as crud
 from app.core.config import settings
 from app.core.security import verify_password
-from app.models import User, UserCreate
+from app.models.user import User, UserCreate
 from app.tests.utils.utils import random_email, random_lower_string
