+++ backend/app/tests/crud/test_user.py
@@ -3,7 +3,7 @@ from sqlmodel import Session
 
 import app.crud.user as crud
 from app.core.security import verify_password
-from app.models import User, UserCreate, UserUpdate
+from app.models.user import User, UserCreate, UserUpdate
 from app.tests.utils.utils import random_email, random_lower_string
