+++ backend/app/tests/utils/item.py
@@ -1,7 +1,7 @@
 from sqlmodel import Session
 
 import app.crud.item as crud
-from app.models import Item, ItemCreate
+from app.models.item import Item, ItemCreate
 from app.tests.utils.user import create_random_user
 from app.tests.utils.utils import random_lower_string
